+++ backend/unified_fake_Shibboleth.php
@@ -35,23 +35,21 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     $email = $_POST['UBIT'] . "@buffalo.edu";
     $id = getInstructorId($con, $email);
     $id_and_name = getStudentInfoFromEmail($con, $email);
-    if (!empty($id)) {
-        $_SESSION['id'] = $id;
-        $_SESSION['redirect'] = 1;
-        http_response_code(302);
-        header("Location: " . FRONTEND_HOME);
-        exit();
-    }
 
-    // Logic for when it is NOT an instructor BUT a student
-    if (!empty($id_and_name)) {
-        session_regenerate_id();
+     // Logic for when it is NOT an instructor BUT a student
+     if (!empty($id_and_name)) {
         $_SESSION['student_id'] = $id_and_name[0];
         $_SESSION['ubit'] = $_POST['UBIT'];
         $_SESSION['redirect'] = 2;
+    }
+    
+    if (!empty($id)) {
+        $_SESSION['id'] = $id;
+        $_SESSION['redirect'] = 1;
+    }
+    if (!empty($_SESSION['redirect'])) {
         http_response_code(302);
         header("Location: " . FRONTEND_HOME);
-        echo implode(',',$_SESSION);
         exit();
     }
